cream demon addition
currently doesnt have special sprites, but this is a modified creambeast. its general stats are increased (speed, vision). its melee converts fat to permanent fat, and giving it a hug bestows the user with a shorter version of the fat curse. currently trying to get it to ignore people who have a fat curse ("blessed")